# api-blank/docs/evidencia/tp2/openapi.json
{"openapi":"3.1.0","info":{"title":"TP2 · Productos, favoritos y listas","description":"Catálogo de DummyJSON de solo lectura. Favoritos y listas persistidos en PostgreSQL, migraciones Flyway y movimiento transaccional de favoritos.","version":"2.0.0"},"servers":[{"url":"http://localhost:8080","description":"Generated server url"}],"tags":[{"name":"Favoritos","description":"CRUD persistido en PostgreSQL; cada favorito pertenece a una lista"},{"name":"Productos","description":"Catálogo externo de solo lectura"},{"name":"Listas","description":"Organización persistente de favoritos"}],"paths":{"/api/favoritos/{id}":{"get":{"tags":["Favoritos"],"summary":"Obtener un favorito","description":"Devuelve los datos y el ID de su lista.","operationId":"obtener","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"Favorito encontrado","content":{"*/*":{"schema":{"$ref":"#/components/schemas/FavoritoResponse"}}}},"404":{"description":"Favorito inexistente","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiError"},"example":{"campos":{},"status":404,"mensaje":"Favorito inexistente"}}}},"409":{"description":"Conflicto de integridad de la lista","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiError"},"example":{"campos":{},"status":409,"mensaje":"Conflicto de integridad de la lista"}}}},"400":{"description":"Datos o ID inválidos","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiError"},"example":{"campos":{},"status":400,"mensaje":"Datos o ID inválidos"}}}}}},"put":{"tags":["Favoritos"],"summary":"Reemplazar los datos de un favorito","description":"Reemplaza productoId, nota y listaId; conserva ID y fecha. No crea si el ID no existe.","operationId":"actualizar","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/FavoritoRequest"}}},"required":true},"responses":{"200":{"description":"Favorito actualizado","content":{"*/*":{"schema":{"$ref":"#/components/schemas/FavoritoResponse"}}}},"404":{"description":"Favorito inexistente","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiError"},"example":{"campos":{},"status":404,"mensaje":"Favorito inexistente"}}}},"409":{"description":"Conflicto de integridad de la lista","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiError"},"example":{"campos":{},"status":409,"mensaje":"Conflicto de integridad de la lista"}}}},"400":{"description":"Datos o ID inválidos","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiError"},"example":{"campos":{},"status":400,"mensaje":"Datos o ID inválidos"}}}}}},"delete":{"tags":["Favoritos"],"summary":"Eliminar un favorito","description":"Elimina el registro persistido; devuelve 404 si no existe.","operationId":"eliminar","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"204":{"description":"Favorito eliminado"},"404":{"description":"Favorito inexistente","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiError"},"example":{"campos":{},"status":404,"mensaje":"Favorito inexistente"}}}},"409":{"description":"Conflicto de integridad de la lista","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiError"},"example":{"campos":{},"status":409,"mensaje":"Conflicto de integridad de la lista"}}}},"400":{"description":"Datos o ID inválidos","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiError"},"example":{"campos":{},"status":400,"mensaje":"Datos o ID inválidos"}}}}}}},"/api/listas":{"get":{"tags":["Listas"],"summary":"Listar listas","description":"Devuelve las listas ordenadas por ID ascendente.","operationId":"listar","responses":{"400":{"description":"Datos inválidos","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiError"},"example":{"campos":{},"status":400,"mensaje":"Datos inválidos"}}}},"404":{"description":"Lista inexistente","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiError"},"example":{"campos":{},"status":404,"mensaje":"Lista inexistente"}}}}}},"post":{"tags":["Listas"],"summary":"Crear una lista","description":"Nombre obligatorio de hasta 100 caracteres.","operationId":"crear","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ListaRequest"}}},"required":true},"responses":{"400":{"description":"Datos inválidos","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiError"},"example":{"campos":{},"status":400,"mensaje":"Datos inválidos"}}}},"404":{"description":"Lista inexistente","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiError"},"example":{"campos":{},"status":404,"mensaje":"Lista inexistente"}}}},"201":{"description":"Lista creada; Location indica su URL","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ListaResponse"}}}}}}},"/api/listas/{origenId}/mover-favoritos":{"post":{"tags":["Listas"],"summary":"Mover favoritos y eliminar origen","description":"Reasigna todos los favoritos a destinoId y elimina el origen en una única transacción. Origen y destino deben ser distintos.","operationId":"mover","parameters":[{"name":"origenId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/MoverFavoritosRequest"}}},"required":true},"responses":{"204":{"description":"Favoritos movidos y origen eliminado"},"409":{"description":"Conflicto de integridad; ninguna escritura queda aplicada","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiError"},"example":{"campos":{},"status":409,"mensaje":"Conflicto de integridad; ninguna escritura queda aplicada"}}}},"400":{"description":"Datos inválidos","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiError"},"example":{"campos":{},"status":400,"mensaje":"Datos inválidos"}}}},"404":{"description":"Lista inexistente","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiError"},"example":{"campos":{},"status":404,"mensaje":"Lista inexistente"}}}}}}},"/api/favoritos":{"get":{"tags":["Favoritos"],"summary":"Listar favoritos","description":"Ordenados por ID ascendente.","operationId":"listar_1","responses":{"409":{"description":"Conflicto de integridad de la lista","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiError"},"example":{"campos":{},"status":409,"mensaje":"Conflicto de integridad de la lista"}}}},"400":{"description":"Datos o ID inválidos","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiError"},"example":{"campos":{},"status":400,"mensaje":"Datos o ID inválidos"}}}}}},"post":{"tags":["Favoritos"],"summary":"Crear un favorito","description":"Genera ID y fecha UTC. Requiere listaId existente (404 si no existe). No verifica la existencia del producto externo.","operationId":"crear_1","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/FavoritoRequest"}}},"required":true},"responses":{"404":{"description":"Lista inexistente","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiError"},"example":{"campos":{},"status":404,"mensaje":"Lista inexistente"}}}},"201":{"description":"Favorito creado; Location indica su URL","content":{"*/*":{"schema":{"$ref":"#/components/schemas/FavoritoResponse"}}}},"409":{"description":"Conflicto de integridad de la lista","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiError"},"example":{"campos":{},"status":409,"mensaje":"Conflicto de integridad de la lista"}}}},"400":{"description":"Datos o ID inválidos","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiError"},"example":{"campos":{},"status":400,"mensaje":"Datos o ID inválidos"}}}}}}},"/health":{"get":{"tags":["health-controller"],"operationId":"health","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object","additionalProperties":{"type":"string"}}}}}}}},"/api/productos":{"get":{"tags":["Productos"],"summary":"Listar productos","description":"Consulta todo el catálogo de DummyJSON y devuelve un DTO propio, sin paginación.","operationId":"listar_2","responses":{"502":{"description":"Proveedor no disponible o respuesta inválida","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiError"},"example":{"campos":{},"status":502,"mensaje":"Proveedor no disponible o respuesta inválida"}}}},"400":{"description":"ID inválido","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiError"},"example":{"campos":{},"status":400,"mensaje":"ID inválido"}}}},"200":{"description":"Catálogo obtenido","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/ProductoResponse"}}}}}}}},"/api/productos/{id}":{"get":{"tags":["Productos"],"summary":"Obtener un producto","description":"Consulta un producto externo por su ID.","operationId":"obtener_1","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"Producto encontrado","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ProductoResponse"}}}},"404":{"description":"Producto inexistente","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiError"},"example":{"campos":{},"status":404,"mensaje":"Producto inexistente"}}}},"502":{"description":"Proveedor no disponible o respuesta inválida","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiError"},"example":{"campos":{},"status":502,"mensaje":"Proveedor no disponible o respuesta inválida"}}}},"400":{"description":"ID inválido","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiError"},"example":{"campos":{},"status":400,"mensaje":"ID inválido"}}}}}}},"/api/listas/{id}":{"get":{"tags":["Listas"],"summary":"Obtener una lista","description":"Devuelve ID y nombre de una lista existente.","operationId":"obtener_2","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"400":{"description":"Datos inválidos","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiError"},"example":{"campos":{},"status":400,"mensaje":"Datos inválidos"}}}},"404":{"description":"Lista inexistente","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiError"},"example":{"campos":{},"status":404,"mensaje":"Lista inexistente"}}}}}},"delete":{"tags":["Listas"],"summary":"Eliminar una lista vacía","description":"Si contiene favoritos devuelve 409 Conflict.","operationId":"eliminar_1","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"204":{"description":"Lista eliminada"},"409":{"description":"La lista todavía tiene favoritos","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiError"},"example":{"campos":{},"status":409,"mensaje":"La lista todavía tiene favoritos"}}}},"400":{"description":"Datos inválidos","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiError"},"example":{"campos":{},"status":400,"mensaje":"Datos inválidos"}}}},"404":{"description":"Lista inexistente","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiError"},"example":{"campos":{},"status":404,"mensaje":"Lista inexistente"}}}}}}},"/api/listas/{id}/favoritos":{"get":{"tags":["Listas"],"summary":"Listar favoritos de una lista","description":"Lista vacía: array vacío. Lista inexistente: 404.","operationId":"favoritos","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"400":{"description":"Datos inválidos","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiError"},"example":{"campos":{},"status":400,"mensaje":"Datos inválidos"}}}},"404":{"description":"Lista inexistente","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiError"},"example":{"campos":{},"status":404,"mensaje":"Lista inexistente"}}}}}}}},"components":{"schemas":{"FavoritoRequest":{"type":"object","properties":{"productoId":{"type":"integer","format":"int64","description":"Referencia a un producto externo; no se consulta al proveedor al guardar","example":1,"minimum":0},"nota":{"type":"string","description":"Nota opcional, hasta 500 caracteres; PUT la reemplaza","example":"Comprar para regalar","maxLength":500,"minLength":0},"listaId":{"type":"integer","format":"int64","description":"ID de una lista existente","example":1,"minimum":0}},"required":["listaId","productoId"]},"ApiError":{"type":"object","properties":{"status":{"type":"integer","format":"int32","example":400},"mensaje":{"type":"string","example":"Hay datos inválidos"},"campos":{"type":"object","additionalProperties":{"type":"string"},"description":"Detalle por campo; vacío si el error no corresponde a campos"}}},"FavoritoResponse":{"type":"object","properties":{"id":{"type":"integer","format":"int64","example":1},"productoId":{"type":"integer","format":"int64","example":1},"nota":{"type":"string","example":"Comprar para regalar"},"fechaAgregado":{"type":"string","format":"date-time","description":"Fecha UTC generada al crear; se conserva al actualizar","example":"2026-09-06T12:00:00Z"},"listaId":{"type":"integer","format":"int64","description":"Lista a la que pertenece el favorito","example":1}}},"ListaRequest":{"type":"object","properties":{"nombre":{"type":"string","example":"Regalos","maxLength":100,"minLength":0}},"required":["nombre"]},"ListaResponse":{"type":"object","properties":{"id":{"type":"integer","format":"int64","example":1},"nombre":{"type":"string","example":"Regalos"}}},"MoverFavoritosRequest":{"type":"object","properties":{"destinoId":{"type":"integer","format":"int64","description":"Lista existente que recibirá los favoritos","example":2,"minimum":0}},"required":["destinoId"]},"ProductoResponse":{"type":"object","properties":{"id":{"type":"integer","format":"int64","example":1},"nombre":{"type":"string","example":"Essence Mascara Lash Princess"},"descripcion":{"type":"string","example":"Máscara de pestañas"},"precio":{"type":"number","example":9.99},"categoria":{"type":"string","example":"beauty"}}}}}}